requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.0.1
Pillow==10.1.0